update for mps

+++ dataset_pm25.py
@@ -151,7 +151,7 @@ def get_dataloader(batch_size, device, validindex=0):
         dataset_valid, batch_size=batch_size, num_workers=1, shuffle=False
     )
 
-    scaler = torch.from_numpy(dataset.train_std).to(device).float()
-    mean_scaler = torch.from_numpy(dataset.train_mean).to(device).float()
+    scaler = torch.from_numpy(dataset.train_std).to(device, dtype=torch.float32).float()
+    mean_scaler = torch.from_numpy(dataset.train_mean).to(device, dtype=torch.float32).float()
 
     return train_loader, valid_loader, test_loader, scaler, mean_scaler
